Calculate triplot that sums up automatic aspect/feature importance grouping

This function shows:

plot for aspect_importance or feature_importance with single aspect

tree that shows aspect_importance for every newly expanded aspect

clustering tree.

calculate_triplot(x, ...)

# S3 method for explainer
calculate_triplot(
  x,
  type = c("predict", "model"),
  new_observation = NULL,
  N = 1000,
  clust_method = "complete",
  ...
)

# S3 method for default
calculate_triplot(
  x,
  data,
  y = NULL,
  predict_function = predict,
  type = c("predict", "model"),
  new_observation = NULL,
  N = 1000,
  clust_method = "complete",
  ...
)

# S3 method for triplot
print(x, ...)

model_triplot(x, ...)

predict_triplot(x, ...)

Arguments

x
an explainer created with the DALEX::explain() function or a model to be explained.

...
other parameters

type
if predict aspect_importance is used, if model is used, than feature_importance is calculated

new_observation
selected observation with columns that corresponds to variables used in the model, should be without target variable

N
number of rows to be sampled from data

clust_method
the agglomeration method to be used, see hclust methods

data
dataset, it will be extracted from x if it's an explainer NOTE: Target variable shouldn't be present in the data

y
true labels for data, will be extracted from x if it's an explainer

predict_function
predict function, it will be extracted from x if it's an explainer

Value

triplot object

Examples

library(DALEX)
set.seed(123)
apartments_num <- apartments[,unlist(lapply(apartments, is.numeric))]
apartments_num_lm_model <- lm(m2.price ~ ., data = apartments_num)
apartments_num_new_observation <- apartments_num[30, ]
explainer_apartments <- explain(model = apartments_num_lm_model,
                                data = apartments_num[,-1],
                                y = apartments_num[, 1],
                                verbose = FALSE)
apartments_tri <- calculate_triplot(x = explainer_apartments,
                                    new_observation =
                                      apartments_num_new_observation[-1])